requirements.txt
matplotlib
pyqt5
top2vec[sentence_encoders]
# top2vec[bert]
umap-learn
